Number Guessing Game
This is a C program that allows the user to play a number guessing game. The user inputs the number of digits they want for the game, and the program generates a random number with that many digits. The user then inputs their guesses until they correctly guess the number.

The program splits the randomly generated number into its individual digits and the user's guess is also split into digits for comparison. The program then outputs the number of digits that are in the correct position and the number of digits that are in the wrong position.

If the user correctly guesses the number, the program outputs the total number of tries it took to guess the number and the guesses from smallest to largest.

This program can be useful for educational purposes or as a simple game for entertainment.